rag_bot/requirements.txt
chromadb>=0.4.0
openai>=1.0.0
tiktoken>=0.5.0
PyMuPDF>=1.23.0
rich>=13.0.0
python-dotenv>=1.0.0
